PrePurchase Counseling  +Batjob:    - Implement import servicer applicant from file  +WebService:Permit not sending an ppCase:    - Just update applicant for this case    - Move two fields ApplicantId and ChgLstUserId from ppCase to applicant (web service field)
git-svn-id: http://nt-svn-jira.vdc.csc.com/HPF/Implementation/Source Code@4998 6d9d970a-e046-46de-96ce-fbc9f6c46dd9

+++ HPF.FutureState/HPF.FutureState.WebServices/AgencyWebService.asmx.cs
@@ -124,7 +124,7 @@ namespace HPF.FutureState.WebServices
                         workingInstance = PrePurchaseCaseSetBL.Instance;
                         PrePurchaseCaseDTO ppc = workingInstance.SavePrePurchaseCaseSet(request.PrePurchaseCaseSet);
                         response.PpcId = ppc.PpcId;
-                        response.AgencyId = (int)ppc.AgencyId;
+                        response.AgencyId = ppc.AgencyId;
                         if (workingInstance.WarningMessages != null && workingInstance.WarningMessages.Count != 0)
                         {
                             response.Status = ResponseStatus.Warning;
